# specs/001-enable-rest-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Compound Interest REST API
  version: "1.0.0"
  description: >
    Controller-based REST API that calculates compound interest and surfaces health telemetry.
servers:
  - url: https://{environment}-calc.example.com
    description: Environment-specific base URL
    variables:
      environment:
        default: staging
        enum:
          - dev
          - staging
          - prod
paths:
  /api/v1/growth/contribution:
    post:
      operationId: postContributionGrowth
      summary: Calculate growth with recurring monthly contributions
      tags:
        - Growth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionGrowthRequest'
            examples:
              payrollInvesting:
                summary: Annual compounding with monthly payroll deposits
                value:
                  principal: 10000
                  annualRatePercent: 5.5
                  compoundingCadence: Annual
                  durationYears: 10
                  monthlyContribution: 100
                  clientReference: payroll-001
      responses:
        '200':
          description: Successful calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResult'
              examples:
                success:
                  value:
                    startingPrincipal: 10000
                    annualRatePercent: 5.5
                    compoundingCadence: Annual
                    durationYears: 10
                    monthlyContribution: 100
                    endingBalance: 33381.64
                    currencyDisplay: "$33,381.64"
                    calculationVersion: v1.0
                    traceId: 5f4c1ae40b4d4f6c949742e4115e9a21
                    responseId: 11111111-2222-3333-4444-555555555555
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security: []
  /api/v1/growth/savings:
    post:
      operationId: postSavingsGrowth
      summary: Calculate compound growth for a fixed balance (HYSA/CD)
      tags:
        - Growth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingsGrowthRequest'
            examples:
              hysa:
                summary: Monthly compounding with no additional deposits
                value:
                  principal: 15000
                  annualRatePercent: 4.2
                  compoundingCadence: Monthly
                  durationYears: 3
      responses:
        '200':
          description: Successful calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResult'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      security: []
  /health/ready:
    get:
      operationId: readinessProbe
      summary: Readiness probe
      tags:
        - Diagnostics
      responses:
        '200':
          description: Service ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [Healthy, Degraded, Unhealthy]
                  checks:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        status:
                          type: string
                        detail:
                          type: string
                  version:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
              example:
                status: Healthy
                checks:
                  - name: configuration
                    status: Healthy
                    detail: All required settings present.
                  - name: logging
                    status: Healthy
                    detail: Controller telemetry connected.
                version: 1.0.0
                timestamp: 2025-10-28T10:00:00Z
        '503':
          description: Service not ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [Degraded, Unhealthy]
                  reason:
                    type: string
                  traceId:
                    type: string
              example:
                status: Unhealthy
                reason: Pending configuration for telemetry sink.
                traceId: 6a1d7f9e5ce34704a7c5f7c31e7ed1df
      security: []
components:
  schemas:
    ContributionGrowthRequest:
      type: object
      required:
        - principal
        - annualRatePercent
        - compoundingCadence
        - durationYears
        - monthlyContribution
      properties:
        principal:
          type: number
          format: decimal
          minimum: 0
          maximum: 10000000
        annualRatePercent:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        compoundingCadence:
          type: string
          enum:
            - Annual
            - SemiAnnual
            - Quarterly
            - Monthly
        durationYears:
          type: integer
          minimum: 0
          maximum: 99
        monthlyContribution:
          type: number
          format: decimal
          minimum: 0
          maximum: 1000000000
        clientReference:
          type: string
          maxLength: 64
        requestedAt:
          type: string
          format: date-time
    SavingsGrowthRequest:
      type: object
      required:
        - principal
        - annualRatePercent
        - compoundingCadence
        - durationYears
      properties:
        principal:
          type: number
          format: decimal
          minimum: 0
          maximum: 10000000
        annualRatePercent:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        compoundingCadence:
          type: string
          enum:
            - Annual
            - SemiAnnual
            - Quarterly
            - Monthly
        durationYears:
          type: integer
          minimum: 0
          maximum: 99
        clientReference:
          type: string
          maxLength: 64
        requestedAt:
          type: string
          format: date-time
    CalculationResult:
      type: object
      required:
        - startingPrincipal
        - annualRatePercent
        - compoundingCadence
        - durationYears
        - monthlyContribution
        - endingBalance
        - currencyDisplay
        - calculationVersion
        - traceId
        - responseId
      properties:
        startingPrincipal:
          type: number
          format: decimal
        annualRatePercent:
          type: number
          format: decimal
        compoundingCadence:
          type: string
        durationYears:
          type: integer
        monthlyContribution:
          type: number
          format: decimal
        endingBalance:
          type: number
          format: decimal
        currencyDisplay:
          type: string
        calculationVersion:
          type: string
        traceId:
          type: string
        responseId:
          type: string
          format: uuid
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        traceId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string